refactor(openapi): extract extractErrorMessages to src/lib/openapi (CP-26)
Pure util moved out of src/index.ts bootstrap. Keeps src/plugins/ strictly
for Elysia instances (rubric firmed in CP-1) and reserves src/lib/openapi/
for future OpenAPI-related utilities. No runtime change — same function,
same call site in mapJsonSchema.zod.override.
